2.0 - 许多更新
- 添加Timeline
- 调整UI
- 添加检查点手动动画

By：QTMK

+++ Assets/#Template/[Scripts]/Level/LevelData.cs
@@ -13,7 +13,7 @@ namespace DancingLineFanmade.Level
     {
         public string levelTitle = "标题";
         public AudioClip soundTrack;
-        [MinValue(0)] public int speed = 12;
+        [MinValue(0f)] public float speed = 12f;
         [MinValue(0f)] public float timeScale = 1f;
         public Vector3 gravity = LevelManager.defaultGravity;
         [TableList] public List<SingleColor> colors = new List<SingleColor>();
